css/css-inline/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Inline Layout properties</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<script>
assert_not_inherited('alignment-baseline', 'central');
assert_not_inherited('baseline-shift', '10px');
assert_inherited('dominant-baseline', 'central');
assert_not_inherited('initial-letters', '2 3');
assert_inherited('initial-letters-align', 'hanging');
assert_inherited('initial-letters-wrap', 'grid');
assert_not_inherited('initial-sizing', 'stretch');
assert_not_inherited('vertical-align', '10px');
</script>
</body>
</html>
